wip: refactoring into cli
