+++ pixi.toml
@@ -24,6 +24,7 @@ matplotlib = "*"
 rattler-build = "*"
 pyyaml = "*"
 pytest = "*"
+typer = ">=0.12.3,<0.13"
 
 [pypi-dependencies]
 repror = { path = ".", editable = true }
